get rid of stdi2p package, move everything relivent to common package
