+++ lib/common/data.go
@@ -2,3 +2,9 @@ package common
 
 // the sha256 of some datastructure
 type IdentHash [32]byte
+
+// i2p date time stamp
+type Date [8]byte
+
+// i2p bytestring
+type String []byte
