--- lib/stdi2p/RouterIdentity.go
@@ -1,5 +1,0 @@
-package stdi2p
-
-// Router Identity
-type RouterIdentity struct {
-}
