--- lib/stdi2p/RouterInfo.go
@@ -1,4 +1,0 @@
-package stdi2p
-
-type RouterInfo struct {
-}
