--- lib/stdi2p/common.go
@@ -1,7 +1,0 @@
-package stdi2p
-
-// i2p date time stamp
-type Date [8]byte
-
-// i2p bytestring
-type String []byte
