--- lib/stdi2p/signing.go
@@ -1,1 +1,0 @@
-package stdi2p
