--- lib/stdi2p/su3.go
@@ -1,5 +1,0 @@
-package stdi2p
-
-// an su3 archive
-type SU3 struct {
-}
